@@ -26,7 +26,6 @@
     "core-js": "^3.6.4",
     "material-design-icons": "^3.0.1",
     "rxjs": "^6.5.4",
-    "rxjs-compat": "^6.5.4",
     "tslib": "^1.10.0",
     "zone.js": "~0.9.1"
   },
@@ -50,7 +49,7 @@
     "rxjs-tslint": "^0.1.7",
     "ts-node": "~3.0.2",
     "tslint": "^6.0.0",
-    "typescript": "^2.9.2",
+    "typescript": "^3.5.3",
     "webpack": "^4.41.5"
   }
 }
